% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\docType{class}
\name{CSarbitrary}
\alias{CSarbitrary}
\alias{CSarbitrary-class}
\title{CSarbitrary Class}
\description{
Represents a cross-section with arbitrary geometry for calculating wetted areas.
}
\section{Slots}{

\describe{
\item{\code{x}}{A numeric vector of x-coordinates.}

\item{\code{z}}{A numeric vector of z-coordinates.}

\item{\code{xb_l}}{A numeric or NULL value representing the left boundary.}

\item{\code{xb_r}}{A numeric or NULL value representing the right boundary.}

\item{\code{kSt_B}}{A numeric or NULL value for the base slope.}

\item{\code{kSt_l}}{A numeric or NULL value for the left slope.}

\item{\code{kSt_r}}{A numeric or NULL value for the right slope.}
}}

\examples{
# Define sample cross-section data
x <- c(0, 4, 9, 13)
z <- c(2, 0, 0, 2)
cs <- CSarbitrary(x = x, z = z, xb_l = 4, xb_r = 9, kSt_B = 35, kSt_l = 45, kSt_r = 45)
}
