% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\docType{class}
\name{CScircle-class}
\alias{CScircle-class}
\alias{CScircle}
\title{CScircle Class}
\description{
Defines a cross-section class with circular geometry for
  hydraulic calculations.
}
\section{Slots}{

\describe{
\item{\code{Di}}{Diameter of the pipe [m].}

\item{\code{kSt}}{Roughness of the pipe according to Strickler [m^(1/3)/s].}

\item{\code{ks}}{Roughness of the pipe according to Prandtl-Coolebrook-White [mm]
(SIA 190)}
}}

\examples{
csC <- CScircle(Di = 1, kSt = 75)
csC <- CScircle(Di = 1, ks = 1.5)
}
