% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pressflow.R
\name{d_aequiv}
\alias{d_aequiv}
\title{Equivalent Hydraulic Diameter}
\usage{
d_aequiv(b, h)
}
\arguments{
\item{b}{Width of the rectangle [m].}

\item{h}{Height of the rectangle [m].}
}
\value{
The equivalent hydraulic diameter [m].
}
\description{
Calculates the equivalent hydraulic diameter of a rectangular
cross-section given its width and height.
}
\examples{
d_aequiv(b = 2, h = 1)
}
