% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethCl_CS_Arbitrary.R
\name{flow}
\alias{flow}
\alias{flow,CSarbitrary-method}
\title{Flow}
\usage{
flow(object, h, J, method = "Strickler", ret = "all",
  plot = FALSE)
}
\arguments{
\item{object}{A `CSarbitrary` object.}

\item{h}{Flow depth [m].}

\item{J}{Bottom slope [-].}

\item{method}{Method to calculate roughness. `method = "Strickler"` considers
equal roughness `KSt_B`, and `method = "Einstein"` estimates a mean roughness.}

\item{ret}{Defines the result returned by the function.}

\item{plot}{Logical; if `TRUE`, plots the results.}
}
\value{
A list containing the following hydraulic variables:
\describe{
  \item{Q}{Discharge [m3/s].}
  \item{v}{Flow velocity [m/s].}
  \item{kSt_m}{Mean roughness [m^(1/3)/s] (if `method = "Einstein"`).}
  \item{A}{Wetted area [m2].}
}
}
\description{
Calculates the discharge of a `CSarbitrary` object for
a given flow depth and bottom slope under uniform flow conditions.
}
\examples{
x <- c(0, 4, 9, 13)
z <- c(2, 0, 0, 2)
cs <- CSarbitrary(
  x = x, z = z, xb_l = 4, xb_r = 9,
  kSt_B = 35, kSt_l = 45, kSt_r = 45
)
flow(cs, h = 2, J = 0.0001, method = "Einstein", ret = "Q")
flow(cs, h = 2, J = 0.0001, method = "Einstein", plot=TRUE)
}
