% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethCl_CS_Arbitrary.R
\name{flow_velocity}
\alias{flow_velocity}
\alias{flow_velocity,CSarbitrary-method}
\title{Flow Velocity}
\usage{
flow_velocity(object, h, J, method = "Strickler")
}
\arguments{
\item{object}{A `CSarbitrary` object.}

\item{h}{Flow depth [m].}

\item{J}{Bottom slope  [-].}

\item{method}{method to calculate the rougness. method="Strickler" considers
equal roughness KSt_B, and method="Einstein" estimates a mean roughness.}
}
\value{
Flow velocity [m/s]
}
\description{
Calculates the flow velocity of a `CSarbitrary` object for a
 given water levels and bottom slope.
}
\examples{
x <- c(0, 4, 9, 13)
z <- c(2, 0, 0, 2)
cs <- CSarbitrary(x = x, z = z, xb_l = 4, xb_r = 9, kSt_B = 35,
                  kSt_l = 45, kSt_r = 45)
flow_velocity(cs, h = 1,J = 0.01, method = "Einstein")

}
