% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethCl_CS_Circle.R
\name{par_fill}
\alias{par_fill}
\alias{par_fill,CScircle-method}
\title{Partial Filling Flow Diagram}
\usage{
par_fill(object,J,method="Strickler")
}
\arguments{
\item{object}{A CScircle object.}

\item{J}{Bottom slope [-].}

\item{method}{Method to calculate the roughness. Allowed are "Strickler"
(equal roughness) and "Prandtl-Coolebrook-White".}
}
\value{
Plots of a partial filling diagram of a circular pipe with discharge and flow velocity
}
\description{
Function to generate a plot of partial-filling diagram of
 circular pipe with discharge and flow velocity
}
\examples{
csC <- CScircle(Di = 0.7, ks = 1.5, kSt = 75)
par_fill(csC,J=0.04)
}
