% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethCl_CS_Arbitrary.R
\name{uniform_flow_Qmax}
\alias{uniform_flow_Qmax}
\alias{uniform_flow_Qmax,CSarbitrary-method}
\title{Uniform Flow Qmax}
\usage{
uniform_flow_Qmax(object, J, method = "Strickler", ret = "all", plot = FALSE)
}
\arguments{
\item{object}{A `CSarbitrary` object.}

\item{J}{Bottom slope [-].}

\item{method}{Method to calculate roughness. `method = "Strickler"` considers
equal roughness `KSt_B`, and `method = "Einstein"` estimates a mean roughness.}

\item{ret}{Defines the result returned by the function.}

\item{plot}{Logical; if `TRUE`, plots the results.}
}
\value{
A list containing the following hydraulic variables:
\describe{
  \item{Qmax}{Maximum discharge [m3/s].}
  \item{hmax}{Maximum flow depth [m].}
  \item{v}{Flow velocity [m/s].}
  \item{kSt_m}{Mean roughness [m^(1/3)/s] (if `method = "Einstein"`).}
  \item{A}{Wetted area [m2].}
}
}
\description{
Calculates the maximum uniform flow discharge of a `CSarbitrary`
object for a given bottom slope.
}
\examples{
x <- c(0, 4, 9, 13)
z <- c(2, 0, 0, 2)
cs <- CSarbitrary(
  x = x, z = z, xb_l = 4, xb_r = 9,
  kSt_B = 35, kSt_l = 45, kSt_r = 45
)
uniform_flow_Qmax(cs, J=0.0001, method="Einstein",ret="Qmax")
uniform_flow_Qmax(cs, J=0.0001, method="Einstein",plot=TRUE)
}
