% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gb_SJ.R
\name{Gb_SJ}
\alias{Gb_SJ}
\title{Sediment transport cappacity according to Smart and Jaeggi}
\usage{
Gb_SJ(d30, dm, d90, J, Rs, um, B,tcrit=0.05,rho_s=2650,s=2.65)
}
\arguments{
\item{d30}{d30 of grain size distribution [m]}

\item{dm}{dm of grain size distribution [m]}

\item{d90}{d90 of grain size distribution [m]}

\item{J}{slope [-]}

\item{Rs}{hydraulic radius [m]}

\item{um}{mean flow velocity [m/s]}

\item{B}{bottom width [m]}

\item{tcrit}{critical shear stress}

\item{rho_s}{density of beadload}

\item{s}{density}
}
\value{
Gb sediment transport rate [kg/s]
}
\description{
Calculates the sediment transport cappacity,i.e. solids-volume per time unity. The forumla is valid for slopes between 0.005 and 0.2.
}
\examples{
d30 <- 0.05
dm <- 0.1
d90 <- 0.2
J <- 0.03
Rs <- 1
um <- 2
B <- 3

Gb_SJ(d30=0.05, dm=0.10, d90=0.2, J=0.03, Rs=1, um=2, B=5)
}
