% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gb_SJ.R
\name{Gb_SJ}
\alias{Gb_SJ}
\title{Sediment transport cappacity according to Smart and Jaeggi}
\usage{
Gb_SJ(d30, dm, d90, J, Rs, um, B)
}
\arguments{
\item{name}{description}
}
\value{
Transport capacity
}
\description{
Calculates the sediment transport cappacity,i.e. solids-volume per time unity. The forumla is valid for slopes between 0.005 and 0.2.
}
\examples{
d30 <- 0.05
dm <- 0.1
d90 <- 0.2
J <- 0.03
Rs <- 1
um <- 2
B <- 3

Gb_SJ(d30=0.05, dm=0.10, d90=0.2, J=0.03, Rs=1, um=2, B=5)
}
